+++ src/zcap_py/__init__.py
@@ -43,7 +43,7 @@ from zcap_py.zcap.parser import ZcapParser
 from zcap_py.zcap.target_attenuation import InvocationTargetAttenuator, PathPrefixAttenuator
 from zcap_py.zcap.verifier import DocumentLoader, ProofVerifier, ZcapVerifier
 
-__version__ = "0.7.0"
+__version__ = "0.8.0"
 
 __all__ = [
     "__version__",